Enable copy to work with int64 strides

+++ mlx/backend/metal/kernels/copy.metal
@@ -1,4 +1,4 @@
-// Copyright © 2023 Apple Inc.
+// Copyright © 2023-2024 Apple Inc.
 
 #include "mlx/backend/metal/kernels/bf16.h"
 #include "mlx/backend/metal/kernels/utils.h"
@@ -23,7 +23,7 @@ template <typename T, typename U>
 [[kernel]] void copy_g_nd1(
     device const T* src,
     device U* dst,
-    constant const size_t& src_stride,
+    constant const int64_t& src_stride,
     uint index [[thread_position_in_grid]]) {
   auto src_idx = elem_to_loc_1(index, src_stride);
   dst[index] = static_cast<U>(src[src_idx]);
@@ -33,11 +33,11 @@ template <typename T, typename U>
 [[kernel]] void copy_g_nd2(
     device const T* src,
     device U* dst,
-    constant const size_t src_strides[2],
+    constant const int64_t src_strides[2],
     uint2 index [[thread_position_in_grid]],
     uint2 grid_dim [[threads_per_grid]]) {
   auto src_idx = elem_to_loc_2(index, src_strides);
-  size_t dst_idx = index.x + (size_t)grid_dim.x * index.y;
+  int64_t dst_idx = index.x + (int64_t)grid_dim.x * index.y;
   dst[dst_idx] = static_cast<U>(src[src_idx]);
 }
 
@@ -45,11 +45,11 @@ template <typename T, typename U>
 [[kernel]] void copy_g_nd3(
     device const T* src,
     device U* dst,
-    constant const size_t src_strides[3],
+    constant const int64_t src_strides[3],
     uint3 index [[thread_position_in_grid]],
     uint3 grid_dim [[threads_per_grid]]) {
   auto src_idx = elem_to_loc_3(index, src_strides);
-  size_t dst_idx = index.x + (size_t)grid_dim.x * (index.y + (size_t)grid_dim.y * index.z);
+  int64_t dst_idx = index.x + (int64_t)grid_dim.x * (index.y + (int64_t)grid_dim.y * index.z);
   dst[dst_idx] = static_cast<U>(src[src_idx]);
 }
 
@@ -58,11 +58,11 @@ template <typename T, typename U, int DIM>
     device const T* src,
     device U* dst,
     constant const int src_shape[DIM],
-    constant const size_t src_strides[DIM],
+    constant const int64_t src_strides[DIM],
     uint3 index [[thread_position_in_grid]],
     uint3 grid_dim [[threads_per_grid]]) {
   auto src_idx = elem_to_loc_nd<DIM>(index, src_shape, src_strides);
-  size_t dst_idx = index.x + (size_t)grid_dim.x * (index.y + (size_t)grid_dim.y * index.z);
+  int64_t dst_idx = index.x + (int64_t)grid_dim.x * (index.y + (int64_t)grid_dim.y * index.z);
   dst[dst_idx] = static_cast<U>(src[src_idx]);
 }
 
@@ -71,12 +71,12 @@ template <typename T, typename U>
     device const T* src,
     device U* dst,
     constant const int* src_shape,
-    constant const size_t* src_strides,
+    constant const int64_t* src_strides,
     constant const int& ndim,
     uint3 index [[thread_position_in_grid]],
     uint3 grid_dim [[threads_per_grid]]) {
   auto src_idx = elem_to_loc(index, src_shape, src_strides, ndim);
-  size_t dst_idx = index.x + (size_t)grid_dim.x * (index.y + (size_t)grid_dim.y * index.z);
+  int64_t dst_idx = index.x + (int64_t)grid_dim.x * (index.y + (int64_t)grid_dim.y * index.z);
   dst[dst_idx] = static_cast<U>(src[src_idx]);
 }
 
@@ -84,8 +84,8 @@ template <typename T, typename U>
 [[kernel]] void copy_gg_nd1(
     device const T* src,
     device U* dst,
-    constant const size_t& src_stride,
-    constant const size_t& dst_stride,
+    constant const int64_t& src_stride,
+    constant const int64_t& dst_stride,
     uint index [[thread_position_in_grid]]) {
   auto src_idx = elem_to_loc_1(index, src_stride);
   auto dst_idx = elem_to_loc_1(index, dst_stride);
@@ -96,8 +96,8 @@ template <typename T, typename U>
 [[kernel]] void copy_gg_nd2(
     device const T* src,
     device U* dst,
-    constant const size_t src_strides[2],
-    constant const size_t dst_strides[2],
+    constant const int64_t src_strides[2],
+    constant const int64_t dst_strides[2],
     uint2 index [[thread_position_in_grid]]) {
   auto src_idx = elem_to_loc_2(index, src_strides);
   auto dst_idx = elem_to_loc_2(index, dst_strides);
@@ -108,8 +108,8 @@ template <typename T, typename U>
 [[kernel]] void copy_gg_nd3(
     device const T* src,
     device U* dst,
-    constant const size_t src_strides[3],
-    constant const size_t dst_strides[3],
+    constant const int64_t src_strides[3],
+    constant const int64_t dst_strides[3],
     uint3 index [[thread_position_in_grid]]) {
   auto src_idx = elem_to_loc_3(index, src_strides);
   auto dst_idx = elem_to_loc_3(index, dst_strides);
@@ -121,8 +121,8 @@ template <typename T, typename U, int DIM>
     device const T* src,
     device U* dst,
     constant const int src_shape[DIM],
-    constant const size_t src_strides[DIM],
-    constant const size_t dst_strides[DIM],
+    constant const int64_t src_strides[DIM],
+    constant const int64_t dst_strides[DIM],
     uint3 index [[thread_position_in_grid]]) {
   auto src_idx = elem_to_loc_nd<DIM>(index, src_shape, src_strides);
   auto dst_idx = elem_to_loc_nd<DIM>(index, src_shape, dst_strides);
@@ -134,8 +134,8 @@ template <typename T, typename U>
     device const T* src,
     device U* dst,
     constant const int* src_shape,
-    constant const size_t* src_strides,
-    constant const size_t* dst_strides,
+    constant const int64_t* src_strides,
+    constant const int64_t* dst_strides,
     constant const int& ndim,
     uint3 index [[thread_position_in_grid]]) {
   auto src_idx = elem_to_loc(index, src_shape, src_strides, ndim);
@@ -156,7 +156,7 @@ template <typename T, typename U>
       device const itype* src, \
       device otype* dst, \
       constant const int src_shape[dims], \
-      constant const size_t src_strides[dims], \
+      constant const int64_t src_strides[dims], \
       uint3 index [[thread_position_in_grid]], \
       uint3 grid_dim [[threads_per_grid]]); \
   template [[host_name("g" name "_" #dims)]] \
@@ -164,8 +164,8 @@ template <typename T, typename U>
       device const itype* src, \
       device otype* dst, \
       constant const int src_shape[dims], \
-      constant const size_t src_strides[dims], \
-      constant const size_t dst_strides[dims], \
+      constant const int64_t src_strides[dims], \
+      constant const int64_t dst_strides[dims], \
       uint3 index [[thread_position_in_grid]]);
 
 
@@ -174,42 +174,42 @@ template <typename T, typename U>
   [[kernel]] void copy_g_nd1<itype, otype>( \
       device const itype* src, \
       device otype* dst, \
-      constant const size_t& src_stride, \
+      constant const int64_t& src_stride, \
       uint index [[thread_position_in_grid]]); \
   template [[host_name(name "_2")]] \
   [[kernel]] void copy_g_nd2<itype, otype>( \
       device const itype* src, \
       device otype* dst, \
-      constant const size_t src_strides[2], \
+      constant const int64_t src_strides[2], \
       uint2 index [[thread_position_in_grid]], \
       uint2 grid_dim [[threads_per_grid]]); \
   template [[host_name(name "_3")]] \
   [[kernel]] void copy_g_nd3<itype, otype>( \
       device const itype* src, \
       device otype* dst, \
-      constant const size_t src_strides[3], \
+      constant const int64_t src_strides[3], \
       uint3 index [[thread_position_in_grid]], \
       uint3 grid_dim [[threads_per_grid]]); \
   template [[host_name("g" name "_1")]] \
   [[kernel]] void copy_gg_nd1<itype, otype>( \
       device const itype* src, \
       device otype* dst, \
-      constant const size_t& src_stride, \
-      constant const size_t& dst_stride, \
+      constant const int64_t& src_stride, \
+      constant const int64_t& dst_stride, \
       uint index [[thread_position_in_grid]]); \
   template [[host_name("g" name "_2")]] \
   [[kernel]] void copy_gg_nd2<itype, otype>( \
       device const itype* src, \
       device otype* dst, \
-      constant const size_t src_strides[2], \
-      constant const size_t dst_strides[2], \
+      constant const int64_t src_strides[2], \
+      constant const int64_t dst_strides[2], \
       uint2 index [[thread_position_in_grid]]); \
   template [[host_name("g" name "_3")]] \
   [[kernel]] void copy_gg_nd3<itype, otype>( \
       device const itype* src, \
       device otype* dst, \
-      constant const size_t src_strides[3], \
-      constant const size_t dst_strides[3], \
+      constant const int64_t src_strides[3], \
+      constant const int64_t dst_strides[3], \
       uint3 index [[thread_position_in_grid]]); \
   instantiate_copy_g_dim(name, itype, otype, 4) \
   instantiate_copy_g_dim(name, itype, otype, 5)
@@ -221,7 +221,7 @@ template <typename T, typename U>
       device const itype* src, \
       device otype* dst, \
       constant const int* src_shape, \
-      constant const size_t* src_strides, \
+      constant const int64_t* src_strides, \
       constant const int& ndim, \
       uint3 index [[thread_position_in_grid]], \
       uint3 grid_dim [[threads_per_grid]]); \
@@ -230,8 +230,8 @@ template <typename T, typename U>
       device const itype* src, \
       device otype* dst, \
       constant const int* src_shape, \
-      constant const size_t* src_strides, \
-      constant const size_t* dst_strides, \
+      constant const int64_t* src_strides, \
+      constant const int64_t* dst_strides, \
       constant const int& ndim, \
       uint3 index [[thread_position_in_grid]]);
